Add files via upload
Main objective of this study is to predict whether a patient is affected with heart disease or not using different machine learning algorithms on a qualified dataset and to find out the co-relations between different attributes. Obtaining clear idea of our proposed Machine Learning techniques and to analyse the result by comparing between the results of different data mining techniques.